requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow[azure]
azure-storage-blob
azure-identity
# azure-mgmt-containerapp
scikit-learn
pandas
joblib
docker
xgboost